+++ Content.Server/_DeltaV/Speech/Components/IrishAccentComponent.cs
@@ -1,6 +1,6 @@
-using Content.Server.DeltaV.Speech.EntitySystems;
+using Content.Server._DeltaV.Speech.EntitySystems;
 
-namespace Content.Server.DeltaV.Speech.Components;
+namespace Content.Server._DeltaV.Speech.Components;
 
 // Takes the ES and assigns the system and component to each other
 [RegisterComponent]
