+++ Content.Server/_DeltaV/Speech/Components/ScottishAccentComponent.cs
@@ -1,6 +1,6 @@
-using Content.Server.DeltaV.Speech.EntitySystems;
+using Content.Server._DeltaV.Speech.EntitySystems;
 
-namespace Content.Server.DeltaV.Speech.Components;
+namespace Content.Server._DeltaV.Speech.Components;
 
 [RegisterComponent]
 [Access(typeof(ScottishAccentSystem))]
