+++ Content.Server/_DeltaV/Speech/EntitySystems/ScottishAccentSystem.cs
@@ -1,11 +1,11 @@
-using Content.Server.DeltaV.Speech.Components;
+using Content.Server._DeltaV.Speech.Components;
 using Content.Server.Speech;
 using Content.Server.Speech.EntitySystems;
 using System.Text.RegularExpressions;
 
-namespace Content.Server.DeltaV.Speech.EntitySystems;
+namespace Content.Server._DeltaV.Speech.EntitySystems;
 
-public sealed class IrishAccentSystem : EntitySystem
+public sealed class ScottishAccentSystem : EntitySystem
 {
     [Dependency] private readonly ReplacementAccentSystem _replacement = default!;
 
@@ -13,20 +13,20 @@ public sealed class IrishAccentSystem : EntitySystem
     {
         base.Initialize();
 
-        SubscribeLocalEvent<IrishAccentComponent, AccentGetEvent>(OnAccentGet);
+        SubscribeLocalEvent<ScottishAccentComponent, AccentGetEvent>(OnAccentGet);
     }
 
     // converts left word when typed into the right word. For example typing you becomes ye.
-    public string Accentuate(string message, IrishAccentComponent component)
+    public string Accentuate(string message, ScottishAccentComponent component)
     {
         var msg = message;
 
-        msg = _replacement.ApplyReplacements(msg, "irish");
+        msg = _replacement.ApplyReplacements(msg, "scottish");
 
         return msg;
     }
 
-    private void OnAccentGet(EntityUid uid, IrishAccentComponent component, AccentGetEvent args)
+    private void OnAccentGet(EntityUid uid, ScottishAccentComponent component, AccentGetEvent args)
     {
         args.Message = Accentuate(args.Message, component);
     }
