+++ Resources/Prototypes/_CosmaticDrift/Reagents/medicine.yml
