+++ Resources/Prototypes/_DeltaV/Shaders/chromatic_aberration.yml
@@ -1,4 +1,4 @@
 - type: shader
   id: ChromaticAberration
   kind: source
-  path: "/Textures/DeltaV/Shaders/chromatic_aberration.swsl"
+  path: "/Textures/_DeltaV/Shaders/chromatic_aberration.swsl"
