+++ Resources/Textures/_DeltaV/Clothing/Uniforms/Jumpsuit/kilt.rsi/equipped-INNERCLOTHING.png
