+++ Resources/Textures/_DeltaV/Clothing/Uniforms/Jumpsuit/kilt.rsi/icon.png
