+++ Resources/Textures/_DeltaV/Clothing/Uniforms/Jumpsuit/kilt.rsi/inhand-left.png
