+++ Resources/Textures/_DeltaV/Clothing/Uniforms/Jumpsuit/kilt.rsi/inhand-right.png
